fix(auth): preserve fallback after quota write failures
